improve TransitEventBuffer

+++ include/quill/backend/BackendWorker.h
@@ -243,7 +243,7 @@ private:
 
     // Phase 1:
     // Read all frontend queues and cache the log statements and the metadata as TransitEvents
-    size_t const cached_transit_events_count = _populate_transit_events_from_frontend_queues();
+    uint64_t const cached_transit_events_count = _populate_transit_events_from_frontend_queues();
 
     if (cached_transit_events_count > 0)
     {
@@ -350,7 +350,7 @@ private:
         break;
       }
 
-      size_t const cached_transit_events_count = _populate_transit_events_from_frontend_queues();
+      uint64_t const cached_transit_events_count = _populate_transit_events_from_frontend_queues();
       if (cached_transit_events_count > 0)
       {
         while (!has_pending_events_for_caching_when_transit_event_buffer_empty() &&
@@ -368,7 +368,7 @@ private:
   /**
    * Populates the local transit event buffer
    */
-  QUILL_ATTRIBUTE_HOT size_t _populate_transit_events_from_frontend_queues()
+  QUILL_ATTRIBUTE_HOT uint64_t _populate_transit_events_from_frontend_queues()
   {
     uint64_t const ts_now = _options.log_timestamp_ordering_grace_period.count()
       ? static_cast<uint64_t>((std::chrono::duration_cast<std::chrono::nanoseconds>(
@@ -377,7 +377,7 @@ private:
                                 .count())
       : std::numeric_limits<uint64_t>::max();
 
-    size_t cached_transit_events_count{0};
+    uint64_t cached_transit_events_count{0};
 
     for (ThreadContext* thread_context : _active_thread_contexts_cache)
     {
@@ -406,7 +406,7 @@ private:
    * @return size of the transit_event_buffer
    */
   template <typename TFrontendQueue>
-  QUILL_ATTRIBUTE_HOT uint32_t _read_and_decode_frontend_queue(TFrontendQueue& frontend_queue,
+  QUILL_ATTRIBUTE_HOT uint64_t _read_and_decode_frontend_queue(TFrontendQueue& frontend_queue,
                                                                ThreadContext* thread_context, uint64_t ts_now)
   {
     // Note: The producer commits only complete messages to the queue.
